+++ wxc_sdk/telephony/hg_and_cq.py
@@ -10,7 +10,7 @@ from ..base import ApiModel, webex_id_to_uuid
 from ..base import SafeEnum as Enum
 from ..common import UserBase, AlternateNumber, IdAndName
 
-__all__ = ['HGandCQ', 'Policy', 'Agent', 'AlternateNumberSettings']
+__all__ = ['HGandCQ', 'Policy', 'Agent', 'AlternateNumberSettings', 'CallingLineIdPolicy']
 
 
 class AlternateNumberSettings(ApiModel):